NavBar Naming Fixes
